+++ libbpf-cargo/src/test.rs
@@ -139,17 +139,17 @@ fn test_build_default() {
     let (_dir, proj_dir, cargo_toml) = setup_temp_project();
 
     // No bpf progs yet
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     // Add prog dir
     create_dir(proj_dir.join("src/bpf")).expect("failed to create prog dir");
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     // Add a prog
     let _prog_file =
         File::create(proj_dir.join("src/bpf/prog.bpf.c")).expect("failed to create prog file");
 
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap();
 
     // Validate generated object file
     validate_bpf_o(proj_dir.as_path().join("target/bpf/prog.bpf.o").as_path());
@@ -167,7 +167,7 @@ fn test_build_invalid_prog() {
         File::create(proj_dir.join("src/bpf/prog.bpf.c")).expect("failed to create prog file");
     writeln!(prog_file, "1").expect("write to prog file failed");
 
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
 }
 
 #[test]
@@ -186,14 +186,14 @@ fn test_build_custom() {
         .expect("write to Cargo.toml failed");
 
     // No bpf progs yet
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     // Add a prog
     create_dir(proj_dir.join("src/other_bpf_dir")).expect("failed to create prog dir");
     let _prog_file = File::create(proj_dir.join("src/other_bpf_dir/prog.bpf.c"))
         .expect("failed to create prog file");
 
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap();
 
     // Validate generated object file
     validate_bpf_o(
@@ -223,13 +223,13 @@ fn test_unknown_metadata_section() {
 
     // Add prog dir
     create_dir(proj_dir.join("src/bpf")).expect("failed to create prog dir");
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     // Add a prog
     let _prog_file =
         File::create(proj_dir.join("src/bpf/prog.bpf.c")).expect("failed to create prog file");
 
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap();
 
     // Validate generated object file
     validate_bpf_o(proj_dir.as_path().join("target/bpf/prog.bpf.o").as_path());
@@ -241,15 +241,15 @@ fn test_enforce_file_extension() {
 
     // Add prog dir
     create_dir(proj_dir.join("src/bpf")).expect("failed to create prog dir");
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     let _prog_file = File::create(proj_dir.join("src/bpf/prog_BAD_EXTENSION.c"))
         .expect("failed to create prog file");
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap_err();
 
     let _prog_file_again = File::create(proj_dir.join("src/bpf/prog_GOOD_EXTENSION.bpf.c"))
         .expect("failed to create prog file");
-    build(true, Some(&cargo_toml), None, Vec::new(), true).unwrap();
+    build(true, Some(&cargo_toml), None, Vec::new()).unwrap();
 }
 
 #[test]
@@ -257,7 +257,7 @@ fn test_build_workspace() {
     let (_dir, _, workspace_cargo_toml, proj_one_dir, proj_two_dir) = setup_temp_workspace();
 
     // No bpf progs yet
-    build(true, Some(&workspace_cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&workspace_cargo_toml), None, Vec::new()).unwrap_err();
 
     // Create bpf prog for project one
     create_dir(proj_one_dir.join("src/bpf")).expect("failed to create prog dir");
@@ -269,7 +269,7 @@ fn test_build_workspace() {
     let _prog_file_2 = File::create(proj_two_dir.join("src/bpf/prog2.bpf.c"))
         .expect("failed to create prog file 2");
 
-    build(true, Some(&workspace_cargo_toml), None, Vec::new(), true).unwrap();
+    build(true, Some(&workspace_cargo_toml), None, Vec::new()).unwrap();
 }
 
 #[test]
@@ -286,7 +286,7 @@ fn test_build_workspace_collision() {
     let _prog_file_2 = File::create(proj_two_dir.join("src/bpf/prog.bpf.c"))
         .expect("failed to create prog file 2");
 
-    build(true, Some(&workspace_cargo_toml), None, Vec::new(), true).unwrap_err();
+    build(true, Some(&workspace_cargo_toml), None, Vec::new()).unwrap_err();
 }
 
 #[test]
@@ -305,7 +305,6 @@ fn test_make_basic() {
         Some(&cargo_toml),
         None,
         Vec::new(),
-        true,
         true,
         Vec::new(),
         None,
@@ -343,7 +342,6 @@ fn test_make_workspace() {
         Some(&workspace_cargo_toml),
         None,
         Vec::new(),
-        true,
         true,
         Vec::new(),
         None,
@@ -401,7 +399,6 @@ fn build_rust_project_from_bpf_c_impl(bpf_c: &str, rust: &str, run: bool) {
         Some(&cargo_toml),
         None,
         Vec::new(),
-        true,
         true,
         Vec::new(),
         None,
@@ -1220,7 +1217,7 @@ fn build_btf_mmap(prog_text: &str) -> Mmap {
     add_vmlinux_header(&proj_dir);
 
     // Build the .bpf.o
-    build(true, Some(&cargo_toml), None, Vec::new(), true).expect("failed to compile");
+    build(true, Some(&cargo_toml), None, Vec::new()).expect("failed to compile");
 
     let obj = OpenOptions::new()
         .read(true)